updated tf models to latest mnist official

+++ keras/mnist-v1/model/official/mnist/dataset.py
@@ -89,7 +89,7 @@ def dataset(directory, images_file, labels_file):
 
   def decode_image(image):
     # Normalize from [0, 255] to [0.0, 1.0]
-    image = tf.decode_raw(image, tf.float32)
+    image = tf.decode_raw(image, tf.uint8)
     image = tf.cast(image, tf.float32)
     image = tf.reshape(image, [784])
     return image / 255.0
